pyproject.toml
[tool.poetry]
name = "surf-calendar"
version = "0.1.0"
description = ""
authors = ["Max-Eugène Mamet <max-eugene.mamet@getbluemorpho.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.140.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
